==> R/2.damage_process.Rmd <==
---
title: "scaling_damage"
author: "abbey yatsko"
date: "2023-02-08"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

libraries
```{r}
library(tidyr)
library(dplyr)
library(stringr)
library(ggplot2)
library(smatr)
library(tidyverse)
library(ggpubr)
library(lme4)
library(easystats)
library(multcompView)
library(ggpmisc)
library(rcompanion)
```

functions
```{r}
# function to calculate trace average (combining upper and lower sections)
trace_means <- function(x){
  area_avg <- aggregate(x$area, by=list(x$id), FUN=mean)
  colnames(area_avg) <- c("id", "area_mean")
  linear_avg <- aggregate(x$linear, by=list(x$id), FUN=mean)
  colnames(linear_avg) <- c("id", "linear_mean")
  out <- merge(area_avg, linear_avg, by = "id",
                  all.x = TRUE)

  return(out)
}
```

data load and format
```{r}
# residrill process output data
df <- read.csv("/Users/abbeyyatsko/Desktop/repos/residrill_MS/R/data_out/damage_est_c.csv")
df <- df[c(2:4)]

# split out tree ID to seperate column - extract trace ID from filename
df$id <- str_extract(df$resi_half, '\\b\\w+$')

# take average of upper and lower half to generate tree-level damage estimates
trace_avg <- trace_means(df)

# read in drill metadata from field
dat <- read.csv("/Users/abbeyyatsko/Desktop/repos/residrill_MS/data/residrill_traces.csv")
dat$id <- gsub('[^[:alnum:] ]','',dat$drill_ID)

# meta data for tree diameters 
diameters <- read.csv("/Users/abbeyyatsko/Desktop/repos/residrill_MS/data/tree_metadata.csv")

# add site variable, append to tree ID (replicates of tree ID number need to be differentiated by site)
diameters$tree_site_ID <- paste(diameters$site, diameters$tree_ID, sep = "-")
diameters <- diameters[c("tree_site_ID", "dbh_cm")]
```

merge and format residrill output with drill heights metadata 
```{r}
# join trace averages with metadata
merge <- left_join(trace_avg, dat, by = "id")
# note: 2 traces in 'dat' dropped (<50%, drill overload)

# add site variable, append to tree ID (replicates of tree ID number need to be differentiated by site)
merge$site <- sub("-.*", "", merge$drill_ID)  
merge$tree_site_ID <- paste(merge$site, merge$tree_ID, sep = "-")

merge_resih <- left_join(merge, diameters, by = "tree_site_ID")

# explore differences in linear/area estimate of damage
ggplot(merge_resih, aes(drill_height_cm)) + 
  geom_smooth(aes(y = area_mean, colour = "area estimate"), method = lm, se=F) + 
  geom_smooth(aes(y = linear_mean, colour = "linear estimate"), method = lm, se=F) +
  facet_wrap(~tree_site_ID)+
  theme_classic()
# working with the linear estimate of damage. this is because symmetry of damage is an assumption used to fill in incomplete traces in processing steps

# subset out three myrtaceae species 
myrt <- subset(merge_resih, species!="Larsenaikia ochreata" & species!="Gardenia?")

# define column for termite presence
myrt$termite_pres <- ifelse(myrt$mound == "Y" | myrt$runway == "Y" | myrt$termite == "Y", "Y", "N")

# add wood density (http://db.worldagroforestry.org/)
# Melaleuca stenostachya (genus) 0.7405
# Eucalyptus cullenii (species) 0.977
# Corymbia clarksoniana (genus) 0.8453

myrt$wd <- NA
myrt$wd <- ifelse(myrt$species == "Melaleuca stenostachya", 0.7405, myrt$wd)
myrt$wd <- ifelse(myrt$species == "Eucalyptus cullenii", 0.977, myrt$wd)
myrt$wd <- ifelse(myrt$species == "Corymbia clarksoniana", 0.8453, myrt$wd)

# calculate damage area 
myrt$V_xsection_cm3 <- ((myrt$drill_dbh_cm/2)^2)*(pi)*(1)
myrt$V_xsection_cm3_damaged <- myrt$V_xsection_cm3 * myrt$linear_mean
myrt$V_xsection_cm3_damaged_linear <- ((myrt$drill_dbh_cm * myrt$linear_mean)/2)^2 * pi
myrt$V_xsection_cm3_damaged_area <- myrt$V_xsection_cm3 * myrt$area_mean

# export working df 
# save(myrt, file="/Users/abbeyyatsko/Desktop/repos/residrill_MS/R/Rdata/myrt.RData")
```

derive overall tree damage 
```{r}
# need to make a different tree ID column with site included 
myrt %>%
  group_by(tree_site_ID, site) %>%
    summarise(across(c("area_mean", "linear_mean", "dbh_cm", "wd"), mean)) -> myrt_treeavg

# merge with tree level damage data
tree_level_damage <- read.csv("/Users/abbeyyatsko/Desktop/repos/residrill_MS/data/tree_external_damage.csv")
tree_level_damage_myrt <- left_join(myrt_treeavg, tree_level_damage, by = "tree_site_ID")

# termite presence column 
tree_level_damage_myrt$termite_pres <- ifelse(tree_level_damage_myrt$mound == "Y" | tree_level_damage_myrt$runway == "Y" | tree_level_damage_myrt$termite == "Y", "Y", "N")

# export working df for tree level averages
# save(tree_level_damage_myrt, file="/Users/abbeyyatsko/Desktop/repos/residrill_MS/R/Rdata/tree_level_damage_myrt.RData")
```

SMA to test linear v area method 
```{r}
# residrill area method and summed cookie damage
sma <- sma(myrt$area_mean ~ myrt$linear_mean)
summary(sma)
plot(sma)

# residuals 
plot(sma, which = "residual") 
abline(0, 0)

# test if the slope is equal to 1
slope_test <- slope.test(x = myrt$linear_mean, y = myrt$area_mean, test.value = 1)

# linear and area based methods significantly correlate, therefore can look at just one
# the correlation is not 1:1, as the slope is 1.14 
# as linear_mean increases, area_mean tends to increase proportionally, though the relationship is slightly steeper than a one-to-one ratio
# from the slope_test, the slope is significantly different from 1
```
