==> R/3.figures_models.Rmd <==
---
title: "draft_models"
author: "abbey yatsko"
date: "2023-04-18"
output: html_document
---

libraries
```{r}
library(tidyr)
library(dplyr)
library(stringr)
library(ggplot2)
library(tidyverse)
library(ggpubr)
library(lme4)
library(easystats)
library(multcompView)
library(ggeffects)
library(nlme)
library(stats)
library(car)
library(ggpmisc)
library(glmmTMB)
library(patchwork)
library(emmeans)
library(rstatix)
```

data load 
```{r}
load("/Users/abbeyyatsko/Desktop/repos/residrill_MS/R/Rdata/tree_level_damage_myrt.RData")
load("/Users/abbeyyatsko/Desktop/repos/residrill_MS/R/Rdata/myrt.RData")

agb <- read.csv("/Users/abbeyyatsko/Desktop/repos/residrill_MS/data/site_agb.csv")
pnw_stems <- read.csv("/Users/abbeyyatsko/Desktop/repos/residrill_MS/data/pnw_biomass.csv")
```

data configure
```{r}
# relevel species
myrt$species <- fct_relevel(myrt$species, "Corymbia clarksoniana", "Eucalyptus cullenii", "Melaleuca stenostachya")

# set graphing parameters/colors
v <- c(10, 25, 34)
low <- "#a3dfba"
mid <- '#347aa3'
high <- '#1f1120'

h <- "#A50026"
m <- '#F99858'
l <- '#364B9A'

custom_colors <- c("Corymbia clarksoniana" = "#A50026", "Eucalyptus cullenii" = "#F99858", "Melaleuca stenostachya" = "#364B9A")
colors <- c("Y" = "#a3dfba", "N" = "#347aa3")
custom_colors_wd <- c("0.7405" = "#a3dfba", "0.8453" = "#347aa3", "0.977" = "#1f1120")
```

Figure 4 models 
```{r}
# model 1. relationship between proportion damage and tree species
m.1 <- glmmTMB(area_mean ~ drill_height_cm  + species + (1|tree_site_ID) + (1|site), myrt, family = beta_family(link = "logit"))
#summary(m.1)
#check_model(m.1)
Anova(m.1)
# posthoc test for m.1
emmeans(m.1, pairwise ~ species, adjust = "tukey")

# model 2. relationship between area damage and tree species
m.2 <- lmer(log(V_xsection_cm3_damaged_area) ~ drill_height_cm + species + (1|tree_site_ID), REML = F, data = myrt)
#summary(m.2)
#check_model(m.2)
Anova(m.2)
# posthoc test for m.2
emmeans(m.2, pairwise ~ species, adjust = "tukey")

# model 3. relationship between proportion damage and height in tree by DBH + wd
m.3 <- glmmTMB(area_mean ~ drill_height_cm  + dbh_cm + wd + (1|tree_site_ID) + (1|site), myrt, family = beta_family(link = "logit"))
#summary(m.3)
#check_model(m.3)
Anova(m.3)

# model 4. relationship between area damage and height in tree by DBH + wd
m.4 <- lmer(log(V_xsection_cm3_damaged_area) ~ drill_height_cm + dbh_cm + wd + (1|tree_site_ID) + (1|site), REML = F, data = myrt)
#summary(m.4)
#check_model(m.4)
Anova(m.4)
```

Figure 4 plots (a-b is about species, c-d is about species for vertical positions, e-f is about traits)
```{r}
# a) visualize species differences - proportion
a <- ggplot(myrt, aes(x = species, y = area_mean, fill = species)) + 
  geom_boxplot() +
  geom_point()+ 
  theme_classic() + 
  scale_fill_brewer(palette = "Set3") +  
  xlab("") + 
  ylab("Proportion damaged") +
  theme(legend.position = "none") +
  scale_x_discrete(labels = c("C. clarksoniana", "E. cullenii", "M. stenostachya")) +
  scale_fill_manual(values = custom_colors)

# b) visualize species differences - area
b <- ggplot(myrt, aes(x = species, y = V_xsection_cm3_damaged_area, fill = species)) + 
  geom_boxplot() +
  geom_point()+ 
  theme_classic() + 
  scale_fill_brewer(palette = "Set3") +  
  xlab("") + 
  ylab("Damaged area (cm2)")  +
  theme(legend.position = "none") +
  scale_x_discrete(labels = c("C. clarksoniana", "E. cullenii", "M. stenostachya")) +
  scale_fill_manual(values = custom_colors)

# c) relationship between proportion damage and tree species at different heights
c <- ggplot(myrt, aes(x = drill_height_cm, y = area_mean, color = species)) + 
  geom_jitter() +
  theme_classic() + 
  xlab("Drill height (cm)") + 
  ylab("Proportion damaged") +
  theme(legend.position = "none") +
  scale_color_manual(values = custom_colors)
  
# d) relationship between area damage and tree species at different heights
area <- ggpredict(m.2, terms = c("drill_height_cm", "species")) 
plot(area, add.data = TRUE)+
   labs(x = "Drill height (cm)", y = "Absolute damaged area (cm2)", title = "Predicted absoute area damage") 

d <-ggplot(myrt, aes(x = drill_height_cm, y = V_xsection_cm3_damaged_area, color = species)) +
  geom_jitter() +
  geom_line(
    aes(x = x, y = predicted),
    data = filter(area, group == 'Corymbia clarksoniana'),
    inherit.aes = FALSE,
    color = h
  ) +
  geom_ribbon(
    aes(x = x, ymin = conf.low, ymax = conf.high),
    data = filter(area, group == 'Corymbia clarksoniana'),
    inherit.aes = FALSE,
    alpha = 0.2,
    fill = h
  ) +
  geom_line(
    aes(x = x, y = predicted),
    data = filter(area, group == 'Eucalyptus cullenii'),
    inherit.aes = FALSE,
    color = m
  ) +
  geom_ribbon(
    aes(x = x, ymin = conf.low, ymax = conf.high),
    data = filter(area, group == 'Eucalyptus cullenii'),
    inherit.aes = FALSE,
    alpha = 0.2,
    fill = m
  ) +
  geom_line(
    aes(x = x, y = predicted),
    data = filter(area, group == 'Melaleuca stenostachya'),
    inherit.aes = FALSE,
    color = l
  ) +
  geom_ribbon(
    aes(x = x, ymin = conf.low, ymax = conf.high),
    data = filter(area, group == 'Melaleuca stenostachya'),
    inherit.aes = FALSE,
    alpha = 0.2,
    fill = l
  ) +
  theme_classic() +
  xlab("Drill height (cm)") +
  ylab("Damaged area (cm²)") +
  scale_color_manual(values = custom_colors) +
  theme(legend.position = "none")

# e) 
e <- ggplot(myrt, aes(x = drill_height_cm, y = area_mean, colour = dbh_cm)) + 
  geom_jitter() +
  theme_classic() + 
  xlab("Drill height (cm)") + 
  ylab("Proportion damaged") +
  scale_colour_gradientn(colours = c("#a3dfba", "#347aa3", "#1f1120"))+
  theme(legend.position = "none")


# f) relationship between area damage and height in tree by size 
area <- ggpredict(m.4, terms = c("drill_height_cm [all]", "dbh_cm [v]"), back.transform = TRUE) #%>% plot(add.data = TRUE)

f <- ggplot(myrt, aes(x = drill_height_cm, y = V_xsection_cm3_damaged)) +
  geom_jitter(aes(color = dbh_cm)) +
  scale_colour_gradient2(
    low = low,
    mid = mid,
    midpoint = 30,
    high = high,
    name = 'DBH'
  ) +
  geom_line(
    aes(x = x, y = predicted),
    data = filter(area, group == '10'),
    inherit.aes = FALSE,
    colour = low
  ) +
  geom_ribbon(
    aes(x = x, ymin = conf.low, ymax = conf.high),
    data = filter(area, group == '10'),
    inherit.aes = FALSE,
    alpha = 0.2,
    fill = low
  ) +
  geom_line(
    aes(x = x, y = predicted),
    data = filter(area, group == '25'),
    inherit.aes = FALSE,
    colour = mid
  ) +
  geom_ribbon(
    aes(x = x, ymin = conf.low, ymax = conf.high),
    data = filter(area, group == '25'),
    inherit.aes = FALSE,
    alpha = 0.2,
    fill = mid
  ) +
  geom_line(
    aes(x = x, y = predicted),
    data = filter(area, group == '34'),
    inherit.aes = FALSE,
    colour = high
  ) +
  geom_ribbon(
    aes(x = x, ymin = conf.low, ymax = conf.high),
    data = filter(area, group == '34'),
    inherit.aes = FALSE,
    alpha = 0.2,
    fill = high
  ) +
  theme_classic() +
  xlab("Drill height (cm)") +
  ylab("Damaged area (cm²)") +
  theme(legend.position = "none")

ggarrange(b, a, d, c, f, e, ncol = 2, nrow = 3)
# plot <- (a+b)/(c+d)/(e+f) 
# out <- plot + plot_annotation(tag_levels = "a")
```

Supplementary Figure 2
Tapering - volume damaged using frustrum 
```{r}
# generate list of all trees 
ids <- unique(myrt$tree_site_ID)

# remove STCK-1 from ids
ids <- ids[!ids %in% "STCK-1"]

# SINGLEPOINT damage estimate (from residrill closest to DBH_)
dbh_resi <- myrt[myrt$drill_height_cm > 113 & myrt$drill_height_cm < 136, ]
missing_elements <- ids[!ids %in% dbh_resi$tree_site_ID]

# troubleshoot missing dbh residrill measurements (STCK-1 only 40cm, STCK-10 needs height = 109, STCK-18 repeated 2x)
stck10 <- myrt %>% filter(tree_site_ID == "STCK-10" & drill_height_cm == 109)
dbh_resi <- rbind(dbh_resi, stck10)
dbh_resi <- dbh_resi %>% filter(!(tree_site_ID == "STCK-18" & drill_height_cm == 114))

# merge in linear_mean of dbh measurement for each tree to frustum df
dbh_resi_merge <- dbh_resi[c("tree_site_ID", "drill_ID", "linear_mean", "drill_dbh_cm")]
x <- dbh_resi_merge$drill_ID
myrt$closest_dbh <- ifelse(myrt$drill_ID %in% x, "Y", "N")

# area of damage at dbh 
myrt$dbh_damage_area_cm2 <- ifelse(myrt$closest_dbh == "Y", (((myrt$drill_dbh_cm/2)^2)*pi) * myrt$area_mean, NA)
  
# function that calculates undamaged frustrum volume 
frus_ud <- function(x){
  ID <- unique(x$tree_site_ID)
  d1 <- min(x$drill_height_cm) 
  d4 <- max(x$drill_height_cm)
  L <- d4 - d1
  A1 <- x$V_xsection_cm3[x$drill_height_cm == d1]
  A4 <- x$V_xsection_cm3[x$drill_height_cm == d4]
  V <- (L/3) * (A1 + sqrt(A1 * A4) + A4)
  
  output <- data.frame(tree_site_ID = ID, frus_undamage_cm3 = V)
  return(output)
}

# function that calculates damaged frustrum volume (with taper)
frus_d <- function(x){
  ID <- unique(x$tree_site_ID)
  d1 <- min(x$drill_height_cm) 
  d4 <- max(x$drill_height_cm)
  L <- d4 - d1
  A1 <- x$V_xsection_cm3_damaged_area[x$drill_height_cm == d1]
  A4 <- x$V_xsection_cm3_damaged_area[x$drill_height_cm == d4]
  V <- (L/3) * (A1 + sqrt(A1 * A4) + A4)
  
  output <- data.frame(tree_site_ID = ID, frus_damage_cm3 = V)
  return(output)
}

# column of damage from assuming constant value based on DBH 
x <- myrt %>% filter(tree_site_ID == "STCK-10")
frus_d_nt <- function(x){
  ID <- unique(x$tree_site_ID)
  d1 <- min(x$drill_height_cm)
  d4 <- max(x$drill_height_cm)
  L <- d4 - d1
  A1 <- x$dbh_damage_area_cm2[x$closest_dbh == "Y"]
  V <- L * (A1)

  output <- data.frame(tree_site_ID = ID, frus_damage_cm3_nt = V)
  return(output)
}

# calculate volume of undamaged frustrum for each tree 
frus_v <- lapply(ids, function(id) frus_ud(myrt[myrt$tree_site_ID == id, ]))
frus_v_out <- do.call(rbind, frus_v)  

# calculate volume of damaged frustrum (taper) for each tree 
frus_v_d <- lapply(ids, function(id) frus_d(myrt[myrt$tree_site_ID == id, ]))
frus_v_d_out <- do.call(rbind, frus_v_d)  

# calculate volume of damaged frustrum (no taper) for each tree 
frus_v_d_nt <- lapply(ids, function(id) frus_d_nt(myrt[myrt$tree_site_ID == id, ]))
frus_v_d_nt_out <- do.call(rbind, frus_v_d_nt)  

# merge frustrum volume estimates for damaged and undamaged 
frustrum <- merge(frus_v_out, frus_v_d_out, by = "tree_site_ID")
frustrum <- merge(frustrum, frus_v_d_nt_out, by = "tree_site_ID")

# calculate biomass using wood density
info <- tree_level_damage_myrt[, c("tree_site_ID", "species", "wd", "dbh_cm", "site")]
frustrum <- merge(info, frustrum, by = "tree_site_ID")

# MULTIPOINT damage estimate
frustrum$undamaged_kg <- (frustrum$wd * frustrum$frus_undamage_cm3)/1000
frustrum$mp_damaged_kg <- (frustrum$wd * frustrum$frus_damage_cm3)/1000
frustrum$sp_damaged_kg <- (frustrum$wd * frustrum$frus_damage_cm3_nt)/1000

# proportion of biomass damaged from single point and multi point estimate
frustrum$sp_damaged_prop <- frustrum$sp_damaged_kg/frustrum$undamaged_kg
frustrum$mp_damaged_prop <- frustrum$mp_damaged_kg/frustrum$undamaged_kg
frustrum_agb_long <- pivot_longer(frustrum, cols = sp_damaged_kg:mp_damaged_kg)

# check normality assumptions
differences <- frustrum$sp_damaged_kg - frustrum$mp_damaged_kg
qqnorm(differences)
qqline(differences, col = "blue")
shapiro.test(differences) # breaks normality assumption 

# nonparametric t-test 
t <- wilcox.test(frustrum$sp_damaged_kg, frustrum$mp_damaged_kg, paired = TRUE) # no difference between groups

# visualization 
frustrum_agb_long %>% group_by(name) %>%
  summarise(mean = mean(value), 
            sd = sd(value)) -> df 

ggplot(df, aes(x = name, y = mean, fill = name))+ 
  geom_bar(stat = "identity", position = "dodge")+
  geom_errorbar(aes(ymin = mean - sd, ymax = mean + sd), width = 0.07)+
  theme_classic()+
  ylab("AGB in frustrum (kg)")+
  xlab("")+
  theme(legend.position="none")
```

Figure 5 models 
```{r}
# merge in presence/absence data for termite/fire presence
tree_damage <- tree_level_damage_myrt[c("tree_site_ID", "site", "termite_pres", "mound", "runway", "termite", "fire")]
frustrum <- left_join(frustrum, tree_damage, by = c("tree_site_ID", "site"))

# Damage ~ species
mod.A <- lmer(log(mp_damaged_kg) ~ species + (1|site), REML = F, data = frustrum)
Anova(mod.A)
# check_model(mod.A)
ggpredict(mod.A, terms = c("species"))  %>% plot(add.data = TRUE)

# post hoc test for which species are different - estimated marginal means
emms <- emmeans(mod.A, ~ species)
pairwise_comparisons_tukey <- pairs(emms, adjust = "tukey")
summary(pairwise_comparisons_tukey)

# Damage ~ traits
mod.B <- lmer(log(mp_damaged_kg) ~ dbh_cm + wd + (1|site), REML = F, data = frustrum)
Anova(mod.B)
#check_model(mod.B)
dam_dbh <- ggpredict(mod.B, terms = c("dbh_cm", "wd"), back.transform = TRUE)# %>% plot(add.data = TRUE)

# Damage ~ external factors
mod.C <- lmer(log(mp_damaged_kg) ~ fire*termite_pres + (1|site), data = frustrum)
Anova(mod.C)
#check_model(mod.C)
ggpredict(mod.C, terms = c("termite_pres"))  %>% plot(add.data = TRUE)
```

Figure 5 plots
```{r}
# Damage ~ species
sp <- ggplot(frustrum, aes(x = species, y = mp_damaged_kg, fill = species))+ 
  geom_boxplot()+
  theme_classic()+
  scale_fill_manual(values = custom_colors)+
  scale_x_discrete(labels = c("C. clarksoniana", "E. cullenii", "M. stenostachya")) +
  xlab("") + 
  ylab("Lower stem damaged biomass (kg)")+
  theme(legend.position = "none")

# Damage ~ traits
frustrum$wd <- as.factor(frustrum$wd)
dam_dbh <- ggpredict(mod.B, terms = c("dbh_cm", "wd"), back.transform = TRUE)# %>% plot(add.data = TRUE)

tr <- ggplot(frustrum, aes(x = dbh_cm, y = log(mp_damaged_kg))) +
  geom_jitter(aes(color = wd)) +
  geom_line(
    aes(x = x, y = log(predicted)),
    data = filter(dam_dbh, group == '0.7405'),
    inherit.aes = FALSE,
    colour = low
  ) +
  geom_ribbon(
    aes(x = x, ymin = log(conf.low), ymax = log(conf.high)),
    data = filter(dam_dbh, group == '0.7405'),
    inherit.aes = FALSE,
    alpha = 0.2,
    fill = low
  ) +
  geom_line(
    aes(x = x, y = log(predicted)),
    data = filter(dam_dbh, group == '0.8453'),
    inherit.aes = FALSE,
    colour = mid
  ) +
  geom_ribbon(
    aes(x = x, ymin = log(conf.low), ymax = log(conf.high)),
    data = filter(dam_dbh, group == '0.8453'),
    inherit.aes = FALSE,
    alpha = 0.2,
    fill = mid
  ) +
  geom_line(
    aes(x = x, y = log(predicted)),
    data = filter(dam_dbh, group == '0.977'),
    inherit.aes = FALSE,
    colour = high
  ) +
  geom_ribbon(
    aes(x = x, ymin = log(conf.low), ymax = log(conf.high)),
    data = filter(dam_dbh, group == '0.977'),
    inherit.aes = FALSE,
    alpha = 0.2,
    fill = high
  ) +
  theme_classic() +
  scale_color_manual(values = custom_colors_wd) +
  xlab("Tree DBH (cm)") +
  ylab("Lower stem damaged biomass (kg)") +
  theme(legend.position = "none")


# Damage ~ external factors
term <- ggplot(frustrum, aes(x = termite_pres.x, y = mp_damaged_kg, fill = termite_pres.x))+ 
  geom_boxplot()+
  theme_classic() + 
  xlab("External termite presence") + 
  ylab("Lower stem damaged biomass (kg)")+
  theme(legend.position = "none")+
  scale_fill_manual(values = colors)

fire <- ggplot(frustrum, aes(x = fire.x, y = mp_damaged_kg, fill = fire.x))+ 
  geom_boxplot()+
  theme_classic() + 
  xlab("External fire scarring") + 
  ylab("Lower stem damaged biomass (kg)")+
  theme(legend.position = "none")+
  scale_fill_manual(values = colors) 

ggarrange(sp, tr, term, fire, ncol = 2, nrow = 2)
```

Figure 6 models 
```{r}
# binomial model - define coding 
myrt$termite_num <- ifelse(myrt$termite_pres == "Y", 1, 0)
myrt$mound_num <- ifelse(myrt$mound == "Y", 1, 0)
myrt$runway_num <- ifelse(myrt$runway == "Y", 1, 0)
myrt$termitesampled_num <- ifelse(myrt$termite == "Y", 1, 0)
myrt$fire_num <- ifelse(myrt$fire == "Y", as.integer(1), as.integer(0))

myrt %>%
  group_by(tree_site_ID, species, site) %>%
  distinct(dbh_cm, wd, fire_num, termite_num, mound_num, runway_num, termitesampled_num)-> pp

# external damage ~ species (chi square)
#pp$termite_num <- as.factor(pp$termite_num)
#pp$fire_num <- as.factor(pp$fire_num)

contingency_table_term <- table(pp$species, pp$termite_num)
fisher_test(contingency_table_term)
ft_term <- pairwise_fisher_test(contingency_table_term)
p_values <- c(ft_term$p.adj)  # Add p-values from all comparisons
p.adjust(p_values, method = "fdr") # adjust with fdr test
# There is a significant difference in termite presence across the species. Euc-Mel is significantly different

contingency_table_fire <- table(pp$species, pp$fire_num)
fisher_test(contingency_table_fire)
# There is no significant difference in fire presence across the species

# external damage ~ traits 
term_dbh <- glm(termite_num ~ dbh_cm, data = pp, family = binomial(logit))
#summary(term_dbh)
Anova(term_dbh)
ggpredict(term_dbh, terms = c("dbh_cm"))  %>% plot(add.data = TRUE)

term_wd <- glm(termite_num ~ wd , data = pp, family = binomial("logit"))
#summary(term_wd)
Anova(term_wd)
ggpredict(term_wd, terms = c("wd")) %>% plot(add.data = TRUE)

fire_dbh <- glm(fire_num ~ dbh_cm, data = pp, family = binomial(logit))
#summary(fire_dbh)
Anova(fire_dbh)
ggpredict(fire_dbh, terms = c("dbh_cm"))  %>% plot(add.data = TRUE)

fire_wd <- glm(fire_num ~ wd , data = pp, family = binomial("logit"))
summary(fire_wd)
Anova(fire_wd)
ggpredict(fire_wd, terms = c("wd")) %>% plot(add.data = TRUE)
```

Figure 6 plots
```{r}
# external damage ~ species 
fig5_a <- ggplot(pp, aes(x = species, fill = termite_num))+ 
  geom_bar(position = "dodge")+
  theme_classic()+
  labs(x = "Species", y = "Count", fill = "External Termite Presence") +  
  scale_fill_manual(values = c("0" = mid, "1" = low), labels = c("Absent", "Present")) +
  scale_x_discrete(labels = c("C. clarksoniana", "E. cullenii", "M. stenostachya"))
fig5_b <- ggplot(pp, aes(x = species, fill = fire_num))+ 
  geom_bar(position = "dodge")+
  theme_classic()+
  labs(x = "Species", y = "Count", fill = "External Fire Scarring") +  
  scale_fill_manual(values = c("0" = mid, "1" = low), labels = c("Absent", "Present")) +
  scale_x_discrete(labels = c("C. clarksoniana", "E. cullenii", "M. stenostachya"))

# external damage ~ traits 
term_dbh <- ggplot(pp, aes(x = dbh_cm.x, y =  termite_num))+
  geom_point()+
  geom_smooth(method = "glm", method.args = list(family = "binomial"), se = TRUE, color = "black")+
  theme_classic()+
  xlab("DBH (cm)")+
  ylab("Termite presence")

term_wd <- ggplot(pp, aes(x = wd, y =  termite_num))+
  geom_point()+
  geom_smooth(method = "glm", method.args = list(family = "binomial"), se = TRUE, color = "black")+
  theme_classic()+
  xlab("Wood density (g/cm3)")+
  ylab("Termite presence")

fire_dbh <- ggplot(pp, aes(x = dbh_cm.x, y =  fire_num))+
  geom_point()+
  geom_smooth(method = "glm", method.args = list(family = "binomial"), se = TRUE, color = "black")+
  theme_classic()+
  xlab("DBH (cm)")+
  ylab("Fire scarring")

fire_wd <- ggplot(pp, aes(x = wd, y =  fire_num))+
  geom_point()+
  geom_smooth(method = "glm", method.args = list(family = "binomial"), se = TRUE, color = "black")+
  theme_classic()+
  xlab("Wood density (g/cm3)")+
  ylab("Fire scarring")

fig5_a + fig5_b + term_dbh + term_wd + patchwork::plot_annotation(tag_levels = "a") 
```

Summary data: Site description, stem percentages, and species biomass
```{r}
# station creek biomass summary
stck_agb <- agb[agb$site == "station_creek",]
stck_agb %>% group_by(species) %>%
  summarise(biomass = sum(agb), 
            count = n(), 
            agb_perc = (biomass/15.38937)*100, 
            stem_perc = (count/50)*100) -> stck_out

# pennyweight biomass calculate
pnw_stems$dbh <- pnw_stems$Circ_cm/(2*pi)
pnw_stems$wd <- NA
pnw_stems$wd <- ifelse(pnw_stems$species == "Melaleuca stenostachya", 0.7405, pnw_stems$wd)
pnw_stems$wd <- ifelse(pnw_stems$species == "Eucalyptus cullenii", 0.977, pnw_stems$wd)
pnw_stems$wd <- ifelse(pnw_stems$species == "Terminalia aridicola subsp. chillagoensis", 0.74, pnw_stems$wd)
pnw_stems$wd <- ifelse(pnw_stems$species == "Gardenia vilhelmii", 0.64, pnw_stems$wd)

pnw_stems$agb <- computeAGB(D = pnw_stems$dbh, WD = pnw_stems$wd, coord = c(144.92,-16.57))
  
# pennyweight biomass summary
pnw_stems %>% group_by(species) %>%
  summarise(biomass = sum(agb), 
            count = n(), 
            agb_perc = (biomass/2.324303)*100, 
            stem_perc = (count/49)*100) -> pnw_out
```

Supplementary Figure 3. size ~ species
```{r}
ggplot(frustrum, aes(x = species, y = dbh_cm, fill = species))+
  geom_boxplot()+
  theme_classic()+
  xlab("")+
  ylab("DBH (cm)")+
  theme(legend.position = "none")

# anova to test species differences in dbh 
anova_result <- aov(dbh_cm ~ species, data = frustrum)
summary(anova_result)
TukeyHSD(anova_result)
```

Extra materials
```{r}
# relationship between dbh and wd 
# model
mod3 <- lm(dbh_cm.x ~ wd, data = pp)
summary(mod3)

ggpredict(mod3, terms = c("wd")) %>% plot(add.data = TRUE)

wd <- ggplot(pp,aes(x= wd, y = dbh_cm.x, colour= species))+
  geom_point()+
  geom_smooth(method = "lm")+
  theme_classic()

ggplot(frustrum, aes(x = species, y = wd, fill = species))+
  geom_boxplot()+
  theme_classic()+
  theme(legend.position = "none")

# how damaged are corymbia trees
length(unique(myrt_c$tree_ID))
# filter myrt_c for 115 ≤ drill_height_cm ≤ 125
t <- myrt_c %>% filter(drill_height_cm >= 112 & drill_height_cm <= 136)
mean(t$linear_mean)

# damage by species 
frustrum_agb %>% group_by(species) %>%
  summarise(mean = mean(sp_damaged_kg)) -> mean_damaged
```

Individual stem tapering 
```{r}
myrt %>% group_by(tree_site_ID) %>% 
  summarise(min_height = min(drill_height_cm), 
            max_height = max(drill_height_cm), 
            low_diam = drill_dbh_cm[which.min(drill_height_cm)],
            high_diam = drill_dbh_cm[which.max(drill_height_cm)], 
            tree_dbh = dbh_cm.x[1], 
            species = species[1]) %>%
  pivot_longer(
    cols = c(low_diam, high_diam), 
    names_to = "diameter_type", 
    values_to = "diameter") %>%
  pivot_longer(
    cols = c(min_height, max_height),     # Second pivot for diameter
    names_to = "height_type",
    values_to = "height") -> taper
  
# keep rows that have diameter_type = low_diam, height_type = min_height OR diameter_type = high_diam, height_type = max_height
taper <- taper %>% filter(diameter_type == "low_diam" & height_type == "min_height" | diameter_type == "high_diam" & height_type == "max_height")

# filter out STCK-1 (only one entry)
taper <- taper %>% filter(tree_site_ID != "STCK-1")

# height type to factor
taper$height_type <- as.factor(taper$height_type)

ggplot(taper, aes(x = as.factor(height_type), y = diameter))+
  geom_boxplot()+
  theme_classic()+
  facet_wrap(~species) 

ggplot(taper, aes(x = height, y = diameter, colour = tree_site_ID))+
  geom_point()+
  geom_smooth(method = "lm")+
  theme_classic()+ 
  theme(legend.position = "none")+
  facet_wrap(~species) 

# is there a significant relationship between height and diameter
mod <- lmer(diameter ~ as.factor(height_type) + (1|tree_site_ID), data = taper)
summary(mod)
Anova(mod)

# consider all points rather than just the min and max
ggplot(myrt, aes(x = drill_height_cm, y = drill_dbh_cm, colour = tree_site_ID))+
  geom_point()+
  geom_smooth(method = "lm", se = F)+
  theme_classic()+
  theme(legend.position = "none")+
  facet_wrap(~species) 

mod <- lmer(drill_dbh_cm ~ drill_height_cm + species + (1|tree_site_ID), data = myrt)
summary(mod)
Anova(mod)
```
